add 'compute/size|size2D.js'
